APK-Build vorbereitet: echte Compilerfehler behoben, CI liefert APK
Das :app-Modul wurde erstmals vollstaendig typgeprueft (gegen selbst
geschriebene Compose-/AndroidX-Stubs, da in der Bauumgebung kein Android SDK
verfuegbar ist). Dabei gefundene und behobene Build-Brecher:

- MainActivity rief die Erweiterungsfunktion viewModels() vollqualifiziert auf
- CoreComponents fehlte der Import fuer den by-Delegaten von animateFloatAsState
- CutsceneScreen uebergab delay() einen Int statt eines Long
- AudioEngine hatte gleichzeitig @Inject-Konstruktor und @Provides-Methode,
  was Dagger als doppelte Bindung ablehnt
- jvmToolchain lag innerhalb des android-Blocks statt auf oberster Ebene
- fehlende Opt-ins fuer experimentelle Material-3-APIs

Zusaetzlich: Workflow .github/workflows/apk.yml baut die Debug-APK, laedt sie
als Artefakt hoch und veroeffentlicht sie als GitHub-Release zum Direkt-Download.

+++ app/src/main/kotlin/com/runeveil/saga/presentation/cutscene/CutsceneScreen.kt
@@ -158,7 +158,7 @@ class CutsceneViewModel @Inject constructor(
     private suspend fun waitOrSkip(durationMs: Int) {
         var elapsed = 0
         while (elapsed < durationMs && !skipRequested) {
-            delay(TICK_MS)
+            delay(TICK_MS.toLong())
             elapsed += TICK_MS
         }
     }
